展示的提供者7

+++ xc-item-detail-provider/src/main/java/cn/axy/xc/xcitemdetailprovider/service/Updinvservice.java
@@ -2,5 +2,5 @@ package cn.axy.xc.xcitemdetailprovider.service;
 
 public interface Updinvservice {
     String updmessageinv(Integer mid,Integer pcs);
-    String updskuinv(Integer mid,Integer skuid,Integer pcs);
+    String updskuinv(Integer mid,Long skuid,Integer pcs);
 }
